small update of int to long in binary i/o

+++ src/XSutils.c
@@ -132,10 +132,10 @@ void binary_dump(long n_isotopes, long n_gridpoints, NuclideGridPoint ** nuclide
 {
 	FILE * fp = fopen("XS_data.dat", "wb");
 	// Dump Nuclide Grid Data
-	for( int i = 0; i < n_isotopes; i++ )
+	for( long i = 0; i < n_isotopes; i++ )
 		fwrite(nuclide_grids[i], sizeof(NuclideGridPoint), n_gridpoints, fp);
 	// Dump UEG Data
-	for( int i = 0; i < n_isotopes * n_gridpoints; i++ )
+	for( long i = 0; i < n_isotopes * n_gridpoints; i++ )
 	{
 		// Write energy level
 		fwrite(&energy_grid[i].energy, sizeof(double), 1, fp);
@@ -152,10 +152,10 @@ void binary_read(long n_isotopes, long n_gridpoints, NuclideGridPoint ** nuclide
 	int stat;
 	FILE * fp = fopen("XS_data.dat", "rb");
 	// Read Nuclide Grid Data
-	for( int i = 0; i < n_isotopes; i++ )
+	for( long i = 0; i < n_isotopes; i++ )
 		stat = fread(nuclide_grids[i], sizeof(NuclideGridPoint), n_gridpoints, fp);
 	// Dump UEG Data
-	for( int i = 0; i < n_isotopes * n_gridpoints; i++ )
+	for( long i = 0; i < n_isotopes * n_gridpoints; i++ )
 	{
 		// Write energy level
 		stat = fread(&energy_grid[i].energy, sizeof(double), 1, fp);
